requirements.txt
bencode.py==3.0.1